rename: Xxx.styled.tsx -> Xxx.styled.ts
