+++ molecules/preview/RecentlyModifiedItem.styled.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import * as font from '~/styles/font.styled';
+import * as font from '~/styles/font';
 
 export const Container = styled.div`
   display: flex;
